Moved upload_ay_reviewable_to_sg to utils
