+++ services/shotgrid_common/ayon_shotgrid_hub/update_from_ayon.py
@@ -353,64 +353,3 @@ def remove_sg_entity_from_ayon_event(
             exc_info=True
         )
 
-def upload_ay_reviewable_to_sg(
-    ayon_event: Dict,
-    sg_session: shotgun_api3.Shotgun,
-    ayon_entity_hub: ayon_api.entity_hub.EntityHub,
-    custom_attribs_map: Dict[str, str],
-):
-    ay_version_id = ayon_event["summary"]["versionId"]
-
-    ay_version_entity = ayon_entity_hub.get_or_query_entity_by_id(
-        ay_version_id, ["version"])
-
-    if not ay_version_entity:
-        raise ValueError(
-            "Event has a non existent version entity "
-            f"'{ay_version_id}'"
-        )
-
-    sg_version_id = ay_version_entity.attribs.get("shotgridId")
-    sg_version_type = ay_version_entity.attribs.get("shotgridType")
-
-    if not sg_version_id:
-        raise ValueError(f"Version '{ay_version_id} not yet synched to SG.")
-
-    ay_file_id = ayon_event["summary"]["fileId"]
-    ay_project_name = ayon_event["project"]
-
-    endpoint = f"projects/{ay_project_name}/files/{ay_file_id}"
-
-    response = ayon_api.get(endpoint)
-    with tempfile.TemporaryDirectory() as temp_dir:
-        temp_file_path = os.path.join(
-            temp_dir,
-            ayon_event["summary"]["filename"]
-        )
-        log.debug(f'Creating temp file at: {temp_file_path}')
-        with open(temp_file_path, 'w+b') as temp_file:
-            temp_file.write(response.content)
-
-        sg_session.upload(
-            "Version",
-            sg_version_id,
-            temp_file_path,
-            field_name="sg_uploaded_movie",
-        )
-
-        endpoint = (f"projects/{ay_project_name}/versions/"
-                    f"{ay_version_id}/thumbnail")
-
-        response = ayon_api.get(endpoint)
-        if not response:
-            log.warning(f"No thumbnail for '{ay_version_id}'.")
-            return
-
-        log.debug(f"Creating thumbnail file at: {temp_file_path}")
-        temp_file_path = os.path.join(temp_dir, "thumbnail.jpg")
-        with open(temp_file_path, 'w+b') as temp_file:
-            temp_file.write(response.content)
-
-        sg_session.upload_thumbnail(
-            sg_version_type, sg_version_id, temp_file_path
-        )
